feat(core): Rework Order shipping to support multiple shipping lines
Relates to #580. This commit changes the data model & API of the Order type so that
shipping is stored as multiple ShippingLine entities. For the time being, we assume a
single ShippingLine per order, but in future we can add the ability to deal with multiple
shipping lines per order.

BREAKING CHANGE: The way shipping charges on Orders are represented has been changed - an Order
now contains multiple ShippingLine entities, each of which has a reference to a ShippingMethod.
This will require a database migration with manual queries to preserve existing order data. See
release blog post for details.