feat(core): Collections can control inheritance of filters
Relates to #1382. This commit introduces a new property on the Collection entity & GraphQL type -
`inheritFilters`. When set to `true` (the default), behaviour is the same as before - a Collection's
filters will be composed of its own filters, plus those of all ancestors up to the root.

Setting to `false` allows you to nest a Collection and set up filters which act completely
independently of the parent filters. This allows much more flexibility in setting up category
hierarchies.

BREAKING CHANGE: The new `inheritFilters` property on the Collection entity will require a DB
migration.