feat(core): Implement MoneyStrategy
Relates to #1835.

BREAKING CHANGE: The introduction of the new MoneyStrategy includes a new GraphQL `Money` scalar,
which replaces `Int` used in v1.x. In practice, this is still a `number` type and should not
break any client applications. One point to note is that `Money` is based on the `Float` scalar
and therefore can represent decimal values, allowing fractions of cents to be represented.