feat(core): Add currencyCode to variant price model
Relates to #1691
This change puts in place a data model which will allow us to support multiple currencies per
channel, and will remove the requirement that a new price be set for
each channel that a variant is assigned to.

BREAKING CHANGE: The `Channel.currencyCode` field has been renamed to `defaultCurrencyCode`, and a
new `currencyCode` field has been added to the `ProductVariantPrice` entity. This will require
a database migration with care taken to preserve exiting data.