feat(core): Make Promotion entity translatable, add description
Relates to #1990

BREAKING CHANGE: The Promotion entity is now translatable, which means existing promotions will need
to be migrated to the new DB schema and care taken to preserve the name data. Also the GraphQL
API for creating and updating Promotions, as well as the corresponding PromotionService methods
have changed to take a `translations` array for setting the `name` and `description` in a given
language.